wip: remove old code
